SSR-18 Add admin part: prepare client, move backend rendering function to client
