+++ src/client/index.js
@@ -1,0 +1,3 @@
+import routes from './bootstrap/routes';
+
+export default { routes };
